apiSetUp
The API is all set up with one accelerometer source that makes up one
batch that I called accelerometerTest I think? Basically the key is
enabled and I enabled android to allow internet so we can import
com.xxx.xxxxx.*;
Now we just need to utilize it! (And a database)
But they have easy methods :D